requirements.txt
crewai==0.121
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.3.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.4.2
httpx>=0.25.0
crewai-tools>=0.45.0
tiktoken>=0.5.1
groundx>=2.3.5
mem0ai==0.1.111 